#50 refactor: ActionBar の mode 切替を ProgressModeButton へ分離
progressMode を ActionBar 本体の hook (useActionBar) が保持していたため、
mode 変更のたびに ActionBar 全体が再レンダリングされ、行動決定・reset・
AsyncSampleButton など無関係なボタンまで巻き込んでいた。
AsyncSampleButton と同様に progressMode 購読を専用コンポーネントへ閉じ込め、
ActionBar 本体を無関係化した。
