+++ src/prototypes/time-control/time-control-03/_components/action-bar/_components/progress-mode-button/index.tsx
@@ -1,0 +1,18 @@
+import { Button } from '@/components/ui/button'
+
+import { useProgressModeButton } from './index.hooks'
+
+/**
+ * 進行モード (auto/manual) 全 actor 一括切替ボタン
+ *
+ * - progressMode 購読をここに閉じ込め、`ActionBar` 本体・他ボタンの再レンダリングを避ける
+ */
+export const ProgressModeButton = () => {
+  const { progressMode, toggleProgressMode } = useProgressModeButton()
+
+  return (
+    <Button onClick={toggleProgressMode} type="button" variant="outline">
+      mode: {progressMode}
+    </Button>
+  )
+}
