+++ src/prototypes/time-control/time-control-03/_components/action-bar/_components/progress-mode-button/index.types.ts
@@ -1,0 +1,8 @@
+import { ProgressMode } from '../../../../types'
+
+export type UseProgressModeButtonReturn = {
+  /** 進行モード (auto/manual) */
+  progressMode: ProgressMode
+  /** 進行モード全 actor 一括切替 */
+  toggleProgressMode: () => void
+}
