+++ src/prototypes/time-control/time-control-03/_components/action-bar/index.hooks.ts
@@ -1,4 +1,3 @@
-import { useTimeControl03Computed } from '../../_computed'
 import { useTimeControl03EventDispatcher } from '../../_events'
 
 import { UseActionBarReturn } from './index.types'
@@ -7,8 +6,6 @@ import { UseActionBarReturn } from './index.types'
  * ActionBar の操作ロジック
  */
 export const useActionBar = (): UseActionBarReturn => {
-  const progressMode = useTimeControl03Computed((state) => state.progressMode)
-
   const timeControl03EventDispatcher = useTimeControl03EventDispatcher()
 
   const dispatchDecision: UseActionBarReturn['dispatchDecision'] =
@@ -17,13 +14,8 @@ export const useActionBar = (): UseActionBarReturn => {
   const resetAll: UseActionBarReturn['resetAll'] =
     timeControl03EventDispatcher['TimeControl03-reset-all']
 
-  const toggleProgressMode: UseActionBarReturn['toggleProgressMode'] =
-    timeControl03EventDispatcher['TimeControl03-toggle-progress-mode']
-
   return {
     dispatchDecision,
-    progressMode,
     resetAll,
-    toggleProgressMode,
   }
 }
